updated for 0.0.3 and renamed to overture_song package and made tests pass
